requirements.txt
discord.py
httpx
apscheduler
python-dotenv
psycopg[binary]
redis
duckdb
